Fix stack overflow without tail-call optimization
With the "-O0" option of GCC, the stack overflow might happens in
the current function invocation because of the recursive "impl()"
function call.

This patch reworks the recursion to the iterative instruction
invocation. It is noteworthy that the performance of the dhrystone
without the JIT compilation is reduced by half.