Implement JIT cache table
In recent, we introduced LLVM "indirectBr" to connect the control flow
to reduce the cost of switching between T2C and interpreter. However,
the hardest part was to mark all the possible destinations and compare
the current PC with them, and we also need to ensure the destination
block being translated in the meanwhile. These difficulty made the
performance worse.

This commit replace the aforesaid implementation with JIT cache table,
which providing another way to avoid the state of rv32emu fallbacking
to the interpreter. This table records the PCs and T1C JIT-ed entries.
When T2C reaches "JALR", the corresponding entry can be invoked by
"LLVMBuildCall2()" and no longer need to handle the destination block
of "indirectBr" in LLVM IR. However, because the current T2C calls the
T1C JIT-ed code, the performance improvement is constrained by the
quality of T1C JIT compiler.

The overall elapsed time of "build/fib.O1.elf" decreases from 125s to
75s (-40%).