Boot Linux kernel stably and simplify peripherals init with calloc
There are some not often crashes when booting Linux kernel, and it
seems like causing by memory operations. With this commit, the Linux
kernel can be booted stably. The reason needs to be investigated
further.

But, this commit cause performance down since the multiple consecutive
SW and multiple consecutive LW fusion cannot be leveraged.