Fix incorrect x86 instruction encoding
The 32-bit immediate value that relates to PC should be interpreted as
the unsigned number and be zero-extended to the length of the width of
host register. Others which relate to the memory address or negative
immediate value should be sign-extended to "__SIZEOF_POINTER__".

To keep consistence, the load/store instruction to/from the x0 register
should be explicitly set to zero value.